+++ ui/src/dataExplorer/components/SaveAsVariable.tsx
@@ -4,30 +4,47 @@ import {withRouter, WithRouterProps} from 'react-router'
 import {connect} from 'react-redux'
 
 // Components
-import VariableFormContext from 'src/variables/components/VariableFormContext'
+import VariableForm from 'src/variables/components/VariableForm'
+
+// Utils
+import {createVariable} from 'src/variables/actions'
+import {extractVariablesList} from 'src/variables/selectors'
 
 // Types
 import {AppState} from 'src/types'
 import {getActiveQuery} from 'src/timeMachine/selectors'
+import {IVariable as Variable} from '@influxdata/influx'
 
 interface OwnProps {
   onHideOverlay: () => void
 }
 
-interface StateProps {
-  initialScript?: string
+interface DispatchProps {
+  onCreateVariable: typeof createVariable
 }
 
-type Props = StateProps & OwnProps
+interface StateProps {
+  initialScript?: string
+  variables: Variable[]
+}
+
+type Props = StateProps & DispatchProps & OwnProps
 
 class SaveAsVariable extends PureComponent<Props & WithRouterProps> {
   render() {
-    const {initialScript, onHideOverlay} = this.props
+    const {
+      onHideOverlay,
+      onCreateVariable,
+      initialScript,
+      variables,
+    } = this.props
 
     return (
-      <VariableFormContext
-        initialScript={initialScript}
+      <VariableForm
         onHideOverlay={onHideOverlay}
+        onCreateVariable={onCreateVariable}
+        initialScript={initialScript}
+        variables={variables}
       />
     )
   }
@@ -35,13 +52,19 @@ class SaveAsVariable extends PureComponent<Props & WithRouterProps> {
 
 const mstp = (state: AppState): StateProps => {
   const activeQuery = getActiveQuery(state)
+  const variables = extractVariablesList(state)
 
   return {
     initialScript: activeQuery.text,
+    variables,
   }
 }
 
-export default connect<StateProps, {}, OwnProps>(
+const mdtp = {
+  onCreateVariable: createVariable,
+}
+
+export default connect<StateProps, DispatchProps, OwnProps>(
   mstp,
-  null
+  mdtp
 )(withRouter<Props>(SaveAsVariable))