+++ ui/src/variables/actions/index.ts
@@ -30,14 +30,8 @@ import * as copy from 'src/shared/copy/notifications'
 
 // Types
 import {Dispatch} from 'redux-thunk'
-import {
-  RemoteDataState,
-  VariableTemplate,
-  QueryArguments,
-  MapArguments,
-  CSVArguments,
-} from 'src/types'
-import {GetState, VariableArgumentType} from 'src/types'
+import {RemoteDataState, VariableTemplate} from 'src/types'
+import {GetState} from 'src/types'
 import {IVariable as Variable, ILabel as Label} from '@influxdata/influx'
 import {VariableValuesByID} from 'src/variables/types'
 import {
@@ -45,94 +39,105 @@ import {
   removeVariableLabelFailed,
 } from 'src/shared/copy/notifications'
 
-export type EditorAction =
-  | ReturnType<typeof clearEditor>
-  | ReturnType<typeof updateType>
-  | ReturnType<typeof updateName>
-  | ReturnType<typeof updateQuery>
-  | ReturnType<typeof updateMap>
-  | ReturnType<typeof updateConstant>
-
-export const clearEditor = () => ({
-  type: 'CLEAR_VARIABLE_EDITOR' as 'CLEAR_VARIABLE_EDITOR',
-})
-
-export const updateType = (type: VariableArgumentType) => ({
-  type: 'CHANGE_VARIABLE_EDITOR_TYPE' as 'CHANGE_VARIABLE_EDITOR_TYPE',
-  payload: type,
-})
-
-export const updateName = (name: string) => ({
-  type: 'UPDATE_VARIABLE_EDITOR_NAME' as 'UPDATE_VARIABLE_EDITOR_NAME',
-  payload: name,
-})
-
-export const updateQuery = (arg: QueryArguments) => ({
-  type: 'UPDATE_VARIABLE_EDITOR_QUERY' as 'UPDATE_VARIABLE_EDITOR_QUERY',
-  payload: arg,
-})
-
-export const updateMap = (arg: MapArguments) => ({
-  type: 'UPDATE_VARIABLE_EDITOR_MAP' as 'UPDATE_VARIABLE_EDITOR_MAP',
-  payload: arg,
-})
-
-export const updateConstant = (arg: CSVArguments) => ({
-  type: 'UPDATE_VARIABLE_EDITOR_CONSTANT' as 'UPDATE_VARIABLE_EDITOR_CONSTANT',
-  payload: arg,
-})
-
 export type Action =
-  | ReturnType<typeof setVariables>
-  | ReturnType<typeof setVariable>
-  | ReturnType<typeof removeVariable>
-  | ReturnType<typeof moveVariable>
-  | ReturnType<typeof setValues>
-  | ReturnType<typeof selectValue>
+  | SetVariables
+  | SetVariable
+  | RemoveVariable
+  | MoveVariable
+  | SetValues
+  | SelectValue
 
-const setVariables = (status: RemoteDataState, variables?: Variable[]) => ({
-  type: 'SET_VARIABLES' as 'SET_VARIABLES',
+interface SetVariables {
+  type: 'SET_VARIABLES'
+  payload: {
+    status: RemoteDataState
+    variables?: Variable[]
+  }
+}
+
+const setVariables = (
+  status: RemoteDataState,
+  variables?: Variable[]
+): SetVariables => ({
+  type: 'SET_VARIABLES',
   payload: {status, variables},
 })
+
+interface SetVariable {
+  type: 'SET_VARIABLE'
+  payload: {
+    id: string
+    status: RemoteDataState
+    variable?: Variable
+  }
+}
 
 const setVariable = (
   id: string,
   status: RemoteDataState,
   variable?: Variable
-) => ({
-  type: 'SET_VARIABLE' as 'SET_VARIABLE',
+): SetVariable => ({
+  type: 'SET_VARIABLE',
   payload: {id, status, variable},
 })
 
-const removeVariable = (id: string) => ({
-  type: 'REMOVE_VARIABLE' as 'REMOVE_VARIABLE',
+interface RemoveVariable {
+  type: 'REMOVE_VARIABLE'
+  payload: {id: string}
+}
+
+const removeVariable = (id: string): RemoveVariable => ({
+  type: 'REMOVE_VARIABLE',
   payload: {id},
 })
+
+interface MoveVariable {
+  type: 'MOVE_VARIABLE'
+  payload: {originalIndex: number; newIndex: number; contextID: string}
+}
 
 export const moveVariable = (
   originalIndex: number,
   newIndex: number,
   contextID: string
-) => ({
-  type: 'MOVE_VARIABLE' as 'MOVE_VARIABLE',
+): MoveVariable => ({
+  type: 'MOVE_VARIABLE',
   payload: {originalIndex, newIndex, contextID},
 })
+
+interface SetValues {
+  type: 'SET_VARIABLE_VALUES'
+  payload: {
+    contextID: string
+    status: RemoteDataState
+    values?: VariableValuesByID
+  }
+}
 
 const setValues = (
   contextID: string,
   status: RemoteDataState,
   values?: VariableValuesByID
-) => ({
-  type: 'SET_VARIABLE_VALUES' as 'SET_VARIABLE_VALUES',
+): SetValues => ({
+  type: 'SET_VARIABLE_VALUES',
   payload: {contextID, status, values},
 })
+
+interface SelectValue {
+  type: 'SELECT_VARIABLE_VALUE'
+  payload: {
+    contextID: string
+    variableID: string
+    selectedValue: string
+  }
+}
 
 export const selectValue = (
   contextID: string,
   variableID: string,
   selectedValue: string
-) => ({
-  type: 'SELECT_VARIABLE_VALUE' as 'SELECT_VARIABLE_VALUE',
+): SelectValue => ({
+  type: 'SELECT_VARIABLE_VALUE',
   payload: {contextID, variableID, selectedValue},
 })