+++ ui/src/variables/components/CreateVariableOverlay.tsx
@@ -1,16 +1,37 @@
 // Libraries
 import React, {PureComponent} from 'react'
+import {connect} from 'react-redux'
 import {withRouter, WithRouterProps} from 'react-router'
+
+// Utils
+import {extractVariablesList} from 'src/variables/selectors'
+
+// Actions
+import {createVariable} from 'src/variables/actions'
 
 // Components
 import {Overlay} from '@influxdata/clockface'
-import VariableFormContext from 'src/variables/components/VariableFormContext'
+import VariableForm from 'src/variables/components/VariableForm'
 import GetResources, {ResourceType} from 'src/shared/components/GetResources'
 
-type Props = WithRouterProps
+// Types
+import {AppState} from 'src/types'
+import {IVariable as Variable} from '@influxdata/influx'
+
+interface DispatchProps {
+  onCreateVariable: typeof createVariable
+}
+
+interface StateProps {
+  variables: Variable[]
+}
+
+type Props = DispatchProps & WithRouterProps & StateProps
 
 class CreateVariableOverlay extends PureComponent<Props> {
   public render() {
+    const {onCreateVariable, variables} = this.props
+
     return (
       <GetResources resource={ResourceType.Variables}>
         <Overlay visible={true}>
@@ -20,7 +41,11 @@ class CreateVariableOverlay extends PureComponent<Props> {
               onDismiss={this.handleHideOverlay}
             />
             <Overlay.Body>
-              <VariableFormContext onHideOverlay={this.handleHideOverlay} />
+              <VariableForm
+                variables={variables}
+                onCreateVariable={onCreateVariable}
+                onHideOverlay={this.handleHideOverlay}
+              />
             </Overlay.Body>
           </Overlay.Container>
         </Overlay>
@@ -38,5 +63,17 @@ class CreateVariableOverlay extends PureComponent<Props> {
   }
 }
 
-export {CreateVariableOverlay}
-export default withRouter<{}>(CreateVariableOverlay)
+const mstp = (state: AppState): StateProps => {
+  const variables = extractVariablesList(state)
+
+  return {variables}
+}
+
+const mdtp: DispatchProps = {
+  onCreateVariable: createVariable,
+}
+
+export default connect<StateProps, DispatchProps>(
+  mstp,
+  mdtp
+)(withRouter<{}>(CreateVariableOverlay))