+++ ui/src/variables/selectors/index.tsx
@@ -6,14 +6,9 @@ import {get} from 'lodash'
 import {getVarAssignment} from 'src/variables/utils/getVarAssignment'
 
 // Types
-import {
-  RemoteDataState,
-  MapArguments,
-  QueryArguments,
-  CSVArguments,
-} from 'src/types'
+import {RemoteDataState} from 'src/types'
 import {VariableAssignment} from 'src/types/ast'
-import {AppState, VariableArguments, VariableArgumentType} from 'src/types'
+import {AppState, VariableArguments} from 'src/types'
 import {
   VariableValues,
   VariableValuesByID,
@@ -34,48 +29,6 @@ const extractVariablesListMemoized = memoizeOne(
 
 export const extractVariablesList = (state: AppState): Variable[] => {
   return extractVariablesListMemoized(state.variables.variables)
-}
-
-export const extractVariableEditorName = (state: AppState): string => {
-  return state.variableEditor.name
-}
-
-export const extractVariableEditorType = (
-  state: AppState
-): VariableArgumentType => {
-  return state.variableEditor.selected
-}
-
-export const extractVariableEditorQuery = (state: AppState): QueryArguments => {
-  return (
-    state.variableEditor.argsQuery || {
-      type: 'query',
-      values: {
-        query: '',
-        language: 'flux',
-      },
-    }
-  )
-}
-
-export const extractVariableEditorMap = (state: AppState): MapArguments => {
-  return (
-    state.variableEditor.argsMap || {
-      type: 'map',
-      values: {},
-    }
-  )
-}
-
-export const extractVariableEditorConstant = (
-  state: AppState
-): CSVArguments => {
-  return (
-    state.variableEditor.argsConstant || {
-      type: 'constant',
-      values: [],
-    }
-  )
 }
 
 const getVariablesForDashboardMemoized = memoizeOne(