+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-
 ## v2.0.0-alpha.19 [unreleased]
 
 ### Features
@@ -6,8 +5,7 @@
 
 ### Bug Fixes
 1. [15295](https://github.com/influxdata/influxdb/pull/15295): Ensures users are created with an active status
-1. [15306](https://github.com/influxdata/influxdb/pull/15306): Added missing string values for CacheStatus type
-1. [15348](https://github.com/influxdata/influxdb/pull/15348): Disable saving for threshold check if no threshold selected
+2. [15306](https://github.com/influxdata/influxdb/pull/15306): Added missing string values for CacheStatus type
 
 ## v2.0.0-alpha.18 [2019-09-26]